updated templates
