+++ components/material-ui/headertemplate.js
@@ -1,12 +1,17 @@
 import React from 'react';
 import AppBar from '@material-ui/core/AppBar';
 import CssBaseline from '@material-ui/core/CssBaseline';
+import Link from '@material-ui/core/Link'
 import Grid from '@material-ui/core/Grid';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
-
+import logo1 from '../../public/img/logo1.png';
+import logo2 from '../../public/img/logo2.png';
+import logo3 from '../../public/img/logo3.png';
+import logo4 from '../../public/img/logo4.png';
+import Home from '../Home';
 
 const useStyles = makeStyles((theme) => ({
 
@@ -49,10 +54,10 @@ export default function Header () {
       <CssBaseline />
       <AppBar position="relative">
         <Toolbar>
-          <img src={'https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Ftse2.mm.bing.net%2Fth%3Fid%3DOIP.tQdEguIK8mW2ralZK8sjGwAAAA%26pid%3DApi&f=1'} height='40px' width='40px' className={classes.icon}/>
-          <Typography variant="h6" color="inherit" noWrap>
-            PocketAlgo
-          </Typography>
+          <img src={logo4} height='60px' width='60px' className={classes.icon}/>
+
+          <Link href="/" variant="h6" color="inherit" noWrap> PocketAlgo</Link>
+
         </Toolbar>
       </AppBar>
