+++ package.json
@@ -39,6 +39,7 @@
     "axios": "^0.19.2",
     "chalk": "^4.1.0",
     "express": "^4.17.1",
+    "file-loader": "^6.0.0",
     "history": "^5.0.0",
     "nodemon": "^2.0.4",
     "react": "^16.13.1",
