+++ webpack.config.js
@@ -25,7 +25,11 @@ module.exports = {
                     'style-loader',
                     'css-loader'
                 ]
-            }
+            },
+            {
+                test: /\.(gif|svg|jpg|png)$/,
+                loader: "file-loader",
+              }
         ]
     },
 };
